hackathonBot-main/requirements.txt
python-telegram-bot==21.6
openai>=1.0.0,<2.0.0
